+++ src/client.ts
@@ -191,7 +191,6 @@ import type {
   SegmentTargetsResponse,
   SegmentType,
   SendFileAPIResponse,
-  SharedLocationRequest,
   SharedLocationResponse,
   SortParam,
   StreamChatOptions,
@@ -209,6 +208,7 @@ import type {
   UpdateChannelTypeResponse,
   UpdateCommandOptions,
   UpdateCommandResponse,
+  UpdateLocationPayload,
   UpdateMessageAPIResponse,
   UpdateMessageOptions,
   UpdatePollAPIResponse,
@@ -233,6 +233,7 @@ import { PollManager } from './poll_manager';
 import type {
   ChannelManagerEventHandlerOverrides,
   ChannelManagerOptions,
+  QueryChannelsRequestType,
 } from './channel_manager';
 import { ChannelManager } from './channel_manager';
 import { NotificationManager } from './notifications';
@@ -720,10 +721,18 @@ export class StreamChat {
   createChannelManager = ({
     eventHandlerOverrides = {},
     options = {},
+    queryChannelsOverride,
   }: {
     eventHandlerOverrides?: ChannelManagerEventHandlerOverrides;
     options?: ChannelManagerOptions;
-  }) => new ChannelManager({ client: this, eventHandlerOverrides, options });
+    queryChannelsOverride?: QueryChannelsRequestType;
+  }) =>
+    new ChannelManager({
+      client: this,
+      eventHandlerOverrides,
+      options,
+      queryChannelsOverride,
+    });
 
   /**
    * Creates a new WebSocket connection with the current user. Returns empty promise, if there is an active connection
@@ -4584,11 +4593,11 @@ export class StreamChat {
   /**
    * updateLocation - Updates a location
    *
-   * @param location UserLocation the location data to update
+   * @param location SharedLocationRequest the location data to update
    *
-   * @returns {Promise<APIResponse>} The server response
+   * @returns {Promise<SharedLocationResponse>} The server response
    */
-  async updateLocation(location: SharedLocationRequest) {
+  async updateLocation(location: UpdateLocationPayload) {
     return await this.put<SharedLocationResponse>(
       this.baseURL + `/users/live_locations`,
       location,