+++ src/messageComposer/messageComposer.ts
@@ -1,14 +1,16 @@
 import { AttachmentManager } from './attachmentManager';
 import { CustomDataManager } from './CustomDataManager';
 import { LinkPreviewsManager } from './linkPreviewsManager';
+import { LocationComposer } from './LocationComposer';
 import { PollComposer } from './pollComposer';
 import { TextComposer } from './textComposer';
-import { DEFAULT_COMPOSER_CONFIG } from './configuration/configuration';
+import { DEFAULT_COMPOSER_CONFIG } from './configuration';
 import type { MessageComposerMiddlewareValue } from './middleware';
 import {
   MessageComposerMiddlewareExecutor,
   MessageDraftComposerMiddlewareExecutor,
 } from './middleware';
+import type { Unsubscribe } from '../store';
 import { StateStore } from '../store';
 import { formatMessage, generateUUIDv4, isLocalMessage, unformatMessage } from '../utils';
 import { mergeWith } from '../utils/mergeWith';
@@ -24,11 +26,11 @@ import type {
   MessageResponse,
   MessageResponseBase,
 } from '../types';
+import { WithSubscriptions } from '../utils/WithSubscriptions';
 import type { StreamChat } from '../client';
 import type { MessageComposerConfig } from './configuration/types';
 import type { DeepPartial } from '../types.utility';
-import type { Unsubscribe } from '../store';
-import { WithSubscriptions } from '../utils/WithSubscriptions';
+import type { MergeWithCustomizer } from '../utils/mergeWith/mergeWithCore';
 
 type UnregisterSubscriptions = Unsubscribe;
 
@@ -129,6 +131,7 @@ export class MessageComposer extends WithSubscriptions {
   linkPreviewsManager: LinkPreviewsManager;
   textComposer: TextComposer;
   pollComposer: PollComposer;
+  locationComposer: LocationComposer;
   customDataManager: CustomDataManager;
   // todo: mediaRecorder: MediaRecorderController;
 
@@ -141,10 +144,6 @@ export class MessageComposer extends WithSubscriptions {
     super();
 
     this.compositionContext = compositionContext;
-
-    this.configState = new StateStore<MessageComposerConfig>(
-      mergeWith(DEFAULT_COMPOSER_CONFIG, config ?? {}),
-    );
 
     // channel is easily inferable from the context
     if (compositionContext instanceof Channel) {
@@ -160,6 +159,32 @@ export class MessageComposer extends WithSubscriptions {
       );
     }
 
+    const mergeChannelConfigCustomizer: MergeWithCustomizer<
+      DeepPartial<MessageComposerConfig>
+    > = (originalVal, channelConfigVal, key) =>
+      typeof originalVal === 'object'
+        ? undefined
+        : originalVal === false && key === 'enabled' // prevent enabling features that are disabled client-side
+          ? false
+          : ['string', 'number', 'bigint', 'boolean', 'symbol'].includes(
+                // prevent enabling features that are disabled server-side
+                typeof channelConfigVal,
+              )
+            ? channelConfigVal // scalar values get overridden by server-side config
+            : originalVal;
+
+    this.configState = new StateStore<MessageComposerConfig>(
+      mergeWith(
+        mergeWith(DEFAULT_COMPOSER_CONFIG, config ?? {}),
+        {
+          location: {
+            enabled: this.channel.getConfig()?.shared_locations,
+          },
+        },
+        mergeChannelConfigCustomizer,
+      ),
+    );
+
     let message: LocalMessage | DraftMessage | undefined = undefined;
     if (compositionIsDraftResponse(composition)) {
       message = composition.message;
@@ -170,6 +195,7 @@ export class MessageComposer extends WithSubscriptions {
 
     this.attachmentManager = new AttachmentManager({ composer: this, message });
     this.linkPreviewsManager = new LinkPreviewsManager({ composer: this, message });
+    this.locationComposer = new LocationComposer({ composer: this, message });
     this.textComposer = new TextComposer({ composer: this, message });
     this.pollComposer = new PollComposer({ composer: this });
     this.customDataManager = new CustomDataManager({ composer: this, message });
@@ -289,7 +315,8 @@ export class MessageComposer extends WithSubscriptions {
       (!this.attachmentManager.uploadsInProgressCount &&
         (!this.textComposer.textIsEmpty ||
           this.attachmentManager.successfulUploadsCount > 0)) ||
-      this.pollId
+      this.pollId ||
+      !!this.locationComposer.validLocation
     );
   }
 
@@ -298,7 +325,8 @@ export class MessageComposer extends WithSubscriptions {
       !this.quotedMessage &&
       this.textComposer.textIsEmpty &&
       !this.attachmentManager.attachments.length &&
-      !this.pollId
+      !this.pollId &&
+      !this.locationComposer.validLocation
     );
   }
 
@@ -320,6 +348,10 @@ export class MessageComposer extends WithSubscriptions {
 
   static generateId = generateUUIDv4;
 
+  refreshId = () => {
+    this.state.partialNext({ id: MessageComposer.generateId() });
+  };
+
   initState = ({
     composition,
   }: { composition?: DraftResponse | MessageResponse | LocalMessage } = {}) => {
@@ -333,6 +365,7 @@ export class MessageComposer extends WithSubscriptions {
           : formatMessage(composition);
     this.attachmentManager.initState({ message });
     this.linkPreviewsManager.initState({ message });
+    this.locationComposer.initState({ message });
     this.textComposer.initState({ message });
     this.pollComposer.initState();
     this.customDataManager.initState({ message });
@@ -403,6 +436,7 @@ export class MessageComposer extends WithSubscriptions {
       this.addUnsubscribeFunction(this.subscribeTextComposerStateChanged());
       this.addUnsubscribeFunction(this.subscribeAttachmentManagerStateChanged());
       this.addUnsubscribeFunction(this.subscribeLinkPreviewsManagerStateChanged());
+      this.addUnsubscribeFunction(this.subscribeLocationComposerStateChanged());
       this.addUnsubscribeFunction(this.subscribePollComposerStateChanged());
       this.addUnsubscribeFunction(this.subscribeCustomDataManagerStateChanged());
       this.addUnsubscribeFunction(this.subscribeMessageComposerStateChanged());
@@ -525,6 +559,18 @@ export class MessageComposer extends WithSubscriptions {
 
   private subscribeAttachmentManagerStateChanged = () =>
     this.attachmentManager.state.subscribe((_, previousValue) => {
+      if (typeof previousValue === 'undefined') return;
+
+      this.logStateUpdateTimestamp();
+
+      if (this.compositionIsEmpty) {
+        this.deleteDraft();
+        return;
+      }
+    });
+
+  private subscribeLocationComposerStateChanged = () =>
+    this.locationComposer.state.subscribe((_, previousValue) => {
       if (typeof previousValue === 'undefined') return;
 
       this.logStateUpdateTimestamp();
@@ -791,6 +837,32 @@ export class MessageComposer extends WithSubscriptions {
         },
         options: {
           type: 'api:poll:create:failed',
+          metadata: {
+            reason: (error as Error).message,
+          },
+          originalError: error instanceof Error ? error : undefined,
+        },
+      });
+      throw error;
+    }
+  };
+
+  sendLocation = async () => {
+    const location = this.locationComposer.validLocation;
+    if (this.threadId || !location) return;
+    try {
+      await this.channel.sendSharedLocation(location);
+      this.refreshId();
+      this.locationComposer.initState();
+    } catch (error) {
+      this.client.notifications.addError({
+        message: 'Failed to share the location',
+        origin: {
+          emitter: 'MessageComposer',
+          context: { composer: this },
+        },
+        options: {
+          type: 'api:location:create:failed',
           metadata: {
             reason: (error as Error).message,
           },